requirements.txt
Brotli==1.0.7
click==7.1.2
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
plotly==4.9.0
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
